+++ src/main/java/org/wso2/integration/com/salesforce/eventbus/protobuf/ManagedFetchRequest.java
