+++ src/main/java/org/wso2/integration/salesforcepubsub/Descriptor.desc
