+++ src/main/java/org/wso2/integration/salesforcepubsub/GRPCConnectionPool.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.wso2.carbon.sfpubsubconnector;
+package org.wso2.integration.salesforcepubsub;
 
 import io.grpc.ManagedChannel;
