+++ src/main/java/org/wso2/integration/salesforcepubsub/TypeConverter.java
@@ -16,7 +16,7 @@
 *  under the License.
 */
 
-package org.wso2.carbon.sfpubsubconnector;
+package org.wso2.integration.salesforcepubsub;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.protobuf.ByteString;
