+++ src/main/resources/functions/publish.xml
@@ -8,7 +8,7 @@
     <!-- Possibly an In mediator or other mediators that read the inbound request -->
         <!-- Initiate Connection -->
 
-    <class name="org.wso2.carbon.sfpubsubconnector.PublishMediator">
+    <class name="org.wso2.integration.salesforcepubsub.PublishMediator">
         <property name="topic_name" expression="$func:topic_name"/>
         <property name="events" expression="$func:events"/>
         <property name="auth_refresh" expression="$func:auth_refresh"/>
